refactor(settings): share ArgsRawInput for launch-args field
Lift raw arg editing into its own shell so ArgsBuilderPage and the gear
shortcut stay aligned without duplicating autocomplete wiring.
